tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-duration_zero_fill_backwards.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-duration: 0s, fill backwards</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-duration-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="A zero-duration animation that fills backwards
   will show the value of the 0% keyframe during any delay period">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move-left {
         from {
            left: 100px;
         }
         to {
            left: 0px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation-name: move-left;
         animation-fill-mode: backwards;
         animation-delay: 1s;
         animation-duration: 0s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         div.className = 'test';
         var passed = (getComputedStyle(div, null).left == '100px');
         test(function(){assert_true(passed);}, testName); done();
         log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should jump leftwards after one second.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
